Rename civis extract/transform modules
