+++ src/dbcp/__init__.py
@@ -10,6 +10,7 @@ import dbcp.etl  # noqa: F401
 import dbcp.extract  # noqa: F401
 import dbcp.extract.acp_projects  # noqa: F401
 import dbcp.extract.ballot_ready  # noqa: F401
+import dbcp.extract.civis  # noqa: F401
 import dbcp.extract.eip_infrastructure  # noqa: F401
 import dbcp.extract.epa_avert  # noqa: F401
 import dbcp.extract.fips_tables  # noqa: F401
@@ -18,7 +19,6 @@ import dbcp.extract.gridstatus_isoqueues  # noqa: F401
 import dbcp.extract.helpers  # noqa: F401
 import dbcp.extract.justice40  # noqa: F401
 import dbcp.extract.lbnl_iso_queue  # noqa: F401
-import dbcp.extract.ljedf  # noqa: F401
 import dbcp.extract.local_opposition  # noqa: F401
 import dbcp.extract.manual_ordinances  # noqa: F401
 import dbcp.extract.ncsl_state_permitting  # noqa: F401
@@ -32,6 +32,7 @@ import dbcp.metadata.data_mart  # noqa: F401
 import dbcp.metadata.data_warehouse  # noqa: F401
 import dbcp.transform.acp_projects  # noqa: F401
 import dbcp.transform.ballot_ready  # noqa: F401
+import dbcp.transform.civis  # noqa: F401
 import dbcp.transform.eip_infrastructure  # noqa: F401
 import dbcp.transform.epa_avert  # noqa: F401
 import dbcp.transform.fips_tables  # noqa: F401
@@ -39,7 +40,6 @@ import dbcp.transform.fyi_queue  # noqa: F401
 import dbcp.transform.gridstatus  # noqa: F401
 import dbcp.transform.justice40  # noqa: F401
 import dbcp.transform.lbnl_iso_queue  # noqa: F401
-import dbcp.transform.ljedf  # noqa: F401
 import dbcp.transform.local_opposition  # noqa: F401
 import dbcp.transform.manual_ordinances  # noqa: F401
 import dbcp.transform.ncsl_state_permitting  # noqa: F401
