+++ src/dbcp/etl.py
@@ -12,8 +12,8 @@ import dbcp
 from dbcp.archivers.utils import ExtractionSettings
 from dbcp.constants import DATA_DIR, OUTPUT_DIR
 from dbcp.extract.ballot_ready import BR_URI
+from dbcp.extract.civis import extract as extract_civis
 from dbcp.extract.fips_tables import CENSUS_URI, TRIBAL_LANDS_URI
-from dbcp.extract.ljedf import extract as extract_civis
 from dbcp.extract.ncsl_state_permitting import NCSLScraper
 from dbcp.helpers import write_to_postgres
 from dbcp.metadata import SchemaName
@@ -176,7 +176,7 @@ def etl_civis() -> dict[str, pd.DataFrame]:
     """ETL archived LJEDF county demographics and election results."""
     raw_dfs = extract_civis()
     county_fips = etl_fips_tables()["census__county_fips"]
-    transformed = dbcp.transform.ljedf.transform(raw_dfs, county_fips=county_fips)
+    transformed = dbcp.transform.civis.transform(raw_dfs, county_fips=county_fips)
     return transformed
